requirements.txt
flask==3.0.2
docker==7.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
GitPython==3.1.42 